Modularize team routes into separate file
improves maintainability and clarity
cleans up index.js
